Perlin noise texture in separate module
