+++ src/texture/mod.rs
@@ -10,36 +10,12 @@ pub use color::*;
 pub mod checker;
 pub use checker::*;
 
+pub mod perlin;
+pub use perlin::*;
+
 pub trait Texture: Debug + Sync + Send {
     fn value(&self, u: f64, v: f64, point: V3) -> Color;
 }
-
-pub struct PerlinTexture {
-    noise: Box<dyn Fn(V3, f64) -> f64 + Sync + Send>,
-    scale: f64,
-}
-
-impl Debug for PerlinTexture {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), Error> {
-        Err(Error)
-    }
-}
-
-impl PerlinTexture {
-    pub fn new(noise: Box<dyn Fn(V3, f64) -> f64 + Sync + Send>, scale: f64) -> PerlinTexture {
-        PerlinTexture { noise, scale }
-    }
-}
-
-impl Texture for PerlinTexture {
-    fn value(&self, u: f64, v: f64, point: V3) -> Color {
-        let noise = (self.noise)(point, self.scale);
-        assert!(noise <= 1.0);
-        assert!(noise >= 0.0);
-        Color(noise * V3::ones())
-    }
-}
-
 
 #[derive(Debug)]
 pub struct ImageTexture {
